fix(retrieve_from_session): incorrectly assuming tokens are stored.
When `store_all_tokens?` is `false` but `require_token_presence_for_authentication?`
is true, the plug helper was performing an extra token resource check
that is already verified by the Jwt.verify/3 function and not required.

This caused an infinite redirect loop back to the sign up page on some
apps.