fix: Add `remember_me` argument to generated magic link sign in action (#1108)
* fix: Add `remember_me` argument to generated magic link sign in action

This is now included by AshAuthenticationPhoenix as of 2.13.0

Uses code as described in https://github.com/team-alembic/ash_authentication_phoenix/issues/692

* fix: add upgrader to add remember_me to magic link sign-in actions

Adds an upgrader function that automatically adds the `remember_me`
argument and `MaybeGenerateTokenChange` change to existing magic link
sign-in actions for users who have both magic_link and remember_me
strategies configured.

This fixes errors in AshAuthenticationPhoenix 2.13.0+ when loading
magic link sign-in pages for apps with remember_me enabled.

* fix: correct assert_has_patch assertion in upgrade test

The `end` keyword was already present in the test file, so it wasn't
being added as a new line in the patch. Removed from the assertion.

---------

Co-authored-by: James Harton <james@harton.nz>